+++ app/code/Codifi/Training/Model/AdminSessionManagement.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 namespace Codifi\Training\Model;
 
 use Magento\Backend\Model\Auth\Session;
+use Magento\Backend\Model\Session as BackendSession;
 
 /**
  * Class AdminSessionManagement
@@ -31,14 +32,24 @@ class AdminSessionManagement
     private $authSession;
 
     /**
+     * Backend session.
+     *
+     * @var BackendSession
+     */
+    private $backendSession;
+
+    /**
      * AdminSessionManagement constructor.
      *
      * @param Session $authSession
+     * @param BackendSession $backendSession
      */
     public function __construct(
-        Session $authSession
+        Session $authSession,
+        BackendSession $backendSession
     ) {
         $this->authSession = $authSession;
+        $this->backendSession = $backendSession;
     }
 
     /**
@@ -51,5 +62,42 @@ class AdminSessionManagement
         $admin = $this->authSession->getUser();
 
         return (int)$admin->getId() ?? 0;
+    }
+
+    /**
+     * Set customer id to array in admin session.
+     *
+     * @param int|null $customerId
+     */
+    public function setCustomerIdToAdminSession(int $customerId = null): void
+    {
+        if (!$customerId) {
+            $customerId = $this->getCustomerId();
+        }
+        $customerIds = $this->getCustomerIds();
+        $customerIds[] = $customerId;
+        $this->authSession->setData(self::ADMIN_SESSION_ATTRIBUTE_CUSTOMER_IDS, $customerIds);
+    }
+
+    /**
+     * Get array customers id from admin session.
+     *
+     * @return array
+     */
+    public function getCustomerIds(): array
+    {
+        return $this->authSession->getData(self::ADMIN_SESSION_ATTRIBUTE_CUSTOMER_IDS) ?? [];
+    }
+
+    /**
+     * Get current customer id from admin session.
+     *
+     * @return int
+     */
+    public function getCustomerId(): int
+    {
+        $customerData = $this->backendSession->getCustomerData();
+
+        return (int)$customerData['account']['id'] ?? 0;
     }
 }
