+++ app/code/Codifi/Training/etc/di.xml
@@ -15,9 +15,6 @@
     </type>
     <type name="Codifi\Training\ViewModel\Admin\AdminSessionManagementViewModel">
         <arguments>
-            <argument name="authSession" xsi:type="object">Magento\Backend\Model\Auth\Session\Proxy</argument>
-        </arguments>
-        <arguments>
             <argument name="backendSession" xsi:type="object">Magento\Backend\Model\Session\Proxy</argument>
         </arguments>
     </type>
